HASKELL QUIZ - DUE WEDNESDAY 3/8
Dear all,
    
Attached here is a short Quiz on Haskell which you have until WEDNESDAY at 11:59PM EST to submit (to your respective GitHub repositories). You are permitted to use your class notes, however if you use the internet/chatGPT I will find out and issue you a score of zero. This is to be done in open book format using class material and textbook ONLY.
        
For ease of reference here is the link to the HASKELL repository:
   
https://github.com/ions29/cpp-reading-material/tree/main/HASKELL