HAIRCUTS.
QUIZ ON CORPORATE, MUNICIPAL, AND GOVERNMENT HAIRCUTS WILL BE RELEASED THIS AFTERNOON.
- DUE BEFORE CLASS IS OVER TOMORROW.

https://www.finra.org/sites/default/files/SEA.Rule_.15c3-1.Interpretations.pdf

GOVERNMENT SECURITIES
(A) GOVERNMENT SECURITIES- PAGE 421
(B) MUNICIPALS- PAGE 441
(E) COMMERCIAL PAPER, BANKERS ACCEPTANCES AND CERTIFICATES OF DEPOSIT - PAGE 461
(F) NONCONVERTIBLE DEBT SECURITIES - PAGE 471




PAGE 157/

/01 Bona Fide Arbitrage
A bona fide arbitrage exists when a long security is convertible into or exchangeable for
a short security, subject to the conditions stated in this subparagraph (J) text, and may, in
lieu of the treatment prescribed in other sections of SEA Rule 15c3-1 be treated as if the
exchange or conversion had been effected, i.e., at realizable values.
For net capital purposes, any cash to be received or paid is credited or charged to net
worth, respectively, in lieu of long or short market values. An example follows.
XYZ preferred is convertible into XYZ common on a share for share basis. XYZ
preferred sells for $120; XYZ common sells for $100. The $20 per share loss, resulting
had the preferred been converted, is charged as a haircut in the computation of net
capital.