https://github.com/ions29/cpp-reading-material/tree/main/XML
                                Section 3, discussion outline attached
                                Cover Letter/ Resume template attached
                                https://github.com/ions29/cpp-reading-material/tree/main/XML 

Dear Class,
         
    
Please review/ update your social media and send a list of links which you believe present a professional outline of yourself. I would also recommend creating a new folder in your GitHub repository for Section 3 as we will have several assignments in this scope.
    
    
The cover letter is due on Friday, April 21st and should be completed using the outline I provided to you in class. For convenience it is also below, the formatting should be used for both your Cover Letter and Resume (which will be due on Monday, April 24th). We will review some Resume/ Interview strategies tomorrow in furtherance of this Section tomorrow.
    
    
Reminder:   Project 1, Part 1 is due today at 11:59PM EST.
    
       
You can use the rest of your time in class today to work on the Project, or your cover letter.
    
    
/s/ BD.
   
        
p.s.    Everyone is now caught up, absent of Jules Martin.
          - can I please have one of the class delegates reach out to him to remind him that his missing assignments are being penalized everyday!?

p.p.s    I attached a copy of the XML file for you all to use as a template for your CV and Resume.
--