Create Instructions
SECTION 2: April 25, 2023
Three (3) page limitation for the Corporate Experience Submission 

https://sam.gov/opp/3221efd1b6044575abb5b1885fbb82fe/view

In no more than three (3) pages, provide sufficient information for the Agency to determine a level of confidence in the ability of the quoter to perform the requirements of the RFQ based on an assessment of relevant experience. 
The quoter shall provide a summary of the quoter’s overall corporate or professional experience performing relevant work. 
    
At a minimum, the quoter shall provide background materials evidencing its knowledge of the industry and data capabilities. Background materials may include descriptions of current data 
systems, testing protocols, staff qualifications, and any other relevant information.